Update README.md
PIZZA SALES ANALYSIS PROJECT REPORT
Introduction to the Project & Problem Statement:
The goal of this project was to analyze and identify insights from daily pizza sales for
the period starting on 1st. January 2015 to 31st. Dec 2015 of an outlet dealing in
the sale of four pizza categories namely; Classic, Chicken, Veggie and Supreme
in five different sizes ie L-Large, M-Medium, S- Regular, XL-X-Large and
XX-Large.
Specifically, the purpose of the project was to compute the following Key Performance
Indicators
1. Total Revenue, as the sum of all revenue generated from the sale of pizzas
2. Average Order Value, as the average revenue generated from an order
3. Total Pizzas Sold, as the sum of all pizza sold throughout the period.
4. Total Orders Placed, as the sum of all orders placed in period of study
5. Average pizzas per order
Identifying any useful trends in the data
1. Daily trend for total Orders
2. Monthly trend for orders
3. Percentage of sales contributed by each pizza category
4. Percentage of sales contributed by each pizza size
5. Total Pizzas sold per category
6. Top 5 and bottom 5 pizza categories by revenue
7. Top 5 and bottom 5 pizza sizes by quantity sold and total orders placed
The above insights and trends will enable decision makers to allocate the business
resources optimally and effectively and maximize revenue and profits